nmstate: Remove kubernetes-nmstate handling code
The kubernetes-nmstate project has now its own operator so users can use
that instead of CNAO.

This change remove the kubernetes-nmstate lifecycle handling code and
also remove the non-managed kubernetes-nmstate from previous
installation if not it will keep the
lock and users will not be able to install standalone kubernetes-nmtate.

It also ensure that only CNAO owned resources are deleted in general,
since ClusterRole and ClusterRoleBindings does not have a namespace.

+++ pkg/apply/apply.go
@@ -87,7 +87,7 @@ func ApplyObject(ctx context.Context, client k8sclient.Client, obj *uns.Unstruct
 }
 
 // DeleteObject deletes an object in the apiserver
-func DeleteObject(ctx context.Context, client k8sclient.Client, obj *uns.Unstructured) error {
+func DeleteOwnedObject(ctx context.Context, client k8sclient.Client, obj *uns.Unstructured) error {
 	name := obj.GetName()
 	namespace := obj.GetNamespace()
 	if name == "" {
@@ -110,11 +110,25 @@ func DeleteObject(ctx context.Context, client k8sclient.Client, obj *uns.Unstruc
 		return nil
 	}
 
+	if !cnaoOwns(existing) {
+		return nil
+	}
+
 	if err := client.Delete(ctx, existing); err != nil {
 		return errors.Wrapf(err, "could not delete %s", objDesc)
 	}
 
 	return nil
+}
+
+func cnaoOwns(obj *uns.Unstructured) bool {
+	owners := obj.GetOwnerReferences()
+	for _, owner := range owners {
+		if owner.Kind == "NetworkAddonsConfig" {
+			return true
+		}
+	}
+	return false
 }
 
 func isTLSSecret(obj *uns.Unstructured) bool {
